[camera][Android] Use downsampling in case of the OOM exception (#20131)
# Why

This a follow-up to the https://exponent-internal.slack.com/archives/C0238THKESC/p1667334481543559.

# How

Added the ability to turn on the downsampling in the case of the OOM exception on Android. Also, when the out-of-memory error is thrown we try to reject the promise instead of crashing the app. However, that may not work - we don't know if the app will work normally after the OOM. 

# Test Plan

- NCL ✅