[camera][web] prevent creating a webworker in non-browser environments (#27222)
# Why

Fix #27194

Expo Router is rendering the page on the server, where the Worker global does not exist.

# How

`createWorkerAsyncFunction` starts a webworker and returns a `decode` function. This function should never be called on the server (it should be from a user event), so when rendering on the server we stub the `decode` function and never create the webworker. 

# Test Plan

<!--
Please describe how you tested this change and how a reviewer could reproduce your test, especially if this PR does not include automated tests! If possible, please also provide terminal output and/or screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff. This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).